Move GetStat() function from stat.go to stats.goto prepare more refactor
